extend stats_t to also compute median and percentiles (need to store all values)

+++ tests/test_core_chrono.cpp
@@ -38,7 +38,7 @@ NANO_CASE(probe)
         NANO_CHECK(probe);
         NANO_CHECK_EQUAL(probe.flops(), flops);
         NANO_CHECK_EQUAL(probe.kflops(), flops / 1024);
-        NANO_CHECK_EQUAL(probe.gflops(), nano::gflops(flops, nano::nanoseconds_t(probe.timings().min())));
+        NANO_CHECK_EQUAL(probe.gflops(), nano::gflops(flops, nano::nanoseconds_t(static_cast<int64_t>(probe.timings().min()))));
 }
 
 NANO_END_MODULE()
